+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 1.11.0-beta.9 (Unreleased)
 -----------------------------
 - Enh #443: Visibility restrictions
+- Fix #444: Fix depreciation warnings on PHP 8.4
 
 1.11.0-beta.8 (July 22, 2025)
 -----------------------------